requirements.txt
pyelftools>=0.29
python-tlsh>=4.7.2; platform_system != "Windows"
